+++ R/run_biomee_f_bysite.R
@@ -343,6 +343,15 @@ run_biomee_f_bysite <- function(
           return(TRUE)
         }
       })
+
+    registered_n_cohorts <- init_cohort$init_n_cohorts[[1]]
+
+    if (registered_n_cohorts != 0) {
+      warning(
+        sprintf("Error: init_cohorts' init_cohort_species is deprecated and should contain a negative value (found %i)",
+                registered_n_cohorts))
+      data_integrity <- append(data_integrity, FALSE)
+    }
     
     # only return true if all checked variables are TRUE
     # suppress warning on coercion of list to single logical